@@ -1,11 +1,11 @@
 import { Pipe, PipeTransform } from '@angular/core';
-import { FileType } from '../../models/filedData';
+import { FileType } from '../../enums';
 
 @Pipe({
     name: 'fileType',
 })
 export class FileTypePipe implements PipeTransform {
     transform(value: FileType): string {
-        return value === 'folder' ? 'File Folder' : 'Text Document';
+        return value === FileType.file ? 'File Folder' : 'Text Document';
     }
 }
